Try to avoid test flakiness in MessagesRpcSpec. (#1657)
The problem summarizes to inserting some messages in the DB, and then
the stream connecting to the DB and reading the existing messages and
*then* receiving notifications about *some* messages being inserted in
the DB, ending in duplicated messages being sent down the stream. So fix
is to sleep a bit after messages are inserted to a avoid this race
condition.

I tried to repro stressing my system but I wasn't successful, so I'm
sending this change hoping I don't see this problem again. I will keep
monitoring and revisit if needed.