Log PrismServerErrors when logging a FutureEither. (#1659)
Many times I have faced that an exception is thrown and I don't know the
root cause, because it gets hidden to the caller. This is OK for
callers, but the server should still log the issue for us to debug and
fix (not just during testing).

After looking at the code closely, I realized we only log exceptions
from a failed `Future`, not caring if the `Either` left was an error
with an exception cause we could also log.

To fix this, I left the wrapping we had, but added a check for
`PrismServerError`s to log its cause.

Note I used an aux function in the fix to make it prettier a bit easier
to read.