Add a testOnlyUntilFailed SBT command to test flaky tests. (#1658)
So I got tired of chasing flaky tests and not having a "built-in" for
it, so came across this way, which I modified to run with `testOnly` as
running a whole sub-project test is way too expensive, but we could add
one if we need to.

The way this works is by simply recursively calling `testOnly` with the
given params passed down, until it fails. For example:

`sbt "project connector" "testOnlyUntilFailed *MessagesRpcSpec -- -z
newer"`

will run the `MessagesRpcSpec` suite, but only the tests that have the `newer` word in it, until a failure occurs.

If you want all tests from the suite, simply do:

`sbt "project connector" "testOnlyUntilFailed *MessagesRpcSpec"`

BTW, I tried an SBT plugin but it was old and it didn't work for us.