Set a default PatienceConfig in AtalaWithPostgresSpec. (#1662)
We most of the time need to set it in the subclasses anyway, and with
very similar values. Just myself, I have added a similar PatienceConfig
to at least 3 tests in the past to avoid test flakiness, and I was about
to, so thought to finally make this change.

Change is big as subclasses needed to be cleaned up. Imports have
cleaned up as well :)