ATA-3007: Create RPC to create a credential issuance. (#1650)
* ATA-3007: Create RPC to create a credential issuance.

Almost all data coming from the client is validated, except for the
contacts belonging to the contacts. This was avoided as it was costly
and unnecessary, as long we validate that actual contacts and groups
belong to the institution.

A DAO getter was added so testing could be done, and this does not
return groups per contacts for the same performance reason as above.
This means that, in the future, editing will be hard, so we may want to
frop the contact-group association we (I mean, me, lol) went with in the
previous PR (schema design). For now, this should be fine, as we only
create an issuance when done, but later we will support stages of it, so
"draft" issuances will be possible.

* Add unimplemented createCredentialIssuance on Connector.

* Move CreateCredentialIssuance RPC to Console.

* Resolve review comments.