circleci: Run publishLocal on the prism-sdk (#1673)
This prevents introducing changes that prevent the modules from being
published, preventing issues like #1670

In order to fix the issues, the java facades from the prism-sdk are removed, also, avoid compiling to scala 2.12. All of this was required to get Android support but the Android app uses the Kotlin SDK now.