Ata 4314 define merkle proofs encoding (#1758)
* ATA-4314: Implement encoding methods for merkle inclusion proofs

This commit implements methods to encode MerkleInclusionProof instances
into strings and back

* ATA-4314: Update console and wallet to use encoded inclusion proofs

This commit updates the models and tables that used merkle inclusion
proofs.
Now, they are encoding them using the methods defined in the SDK

* ATA-4314: Change merkle proofs encoding to JSON

This commit changes the encoding format to a JSON object instead of a
custom string format. The advantage is to allow languages that do not
have our SDK, to parse it with less work. We opted for JSON insteda of
Protobuf to avoid adding the dependency to protobuf in the crypto
library

* ATA-4314: Rename migration V36 to V37

After rebase, there was a V36 migration added from develop.
This commit updates the name of the migration defined in this branch to
avoid collision