ATA-4345 send store merkle proofs (#1788)
* ATA-4345: Update models and backend to send merkle proofs

This commit updates getBy methods to return the merkle inclusion proofs
associated to credentials. It also updates PlainTextCredential model to
send this new proof value

* ATA-4345: Update the wallet and console service

This commit updates the wallet task that copies received credentials in
the connector. The commit also updates the console models and db to
store the newly added inclusion proofs
The commit also do some name changes for consistency

* ATA-4345: Fix tests

* ATA-4345: Minor fix after rebase

* ATA-4345: Fix stored credentials service tests