ATA-5478: Temporal fix for credential revocation
The wallet is unable to revoke credentials because the node requires
a revocation key, unfortunately, there is no simple way to add such
key to the wallet until the SDK is upgraded, which is blocked by a
bug in kotlinjs.

As a temporal workaround, let's allow revoking credentials with the issuance key.