ATA-5562: Add missing foreign key to the received_credentials table
Now, the contact_id has a foreign key to make sure that a received
credential is always linked to a contact.

While the tests succeed, the integration with the wallet is now broken
because the wallet is submitting a connection_id instead of a contact_id,
it is better to be aware of the issue instead of thinking that everything
is working.