ATL-611: Treat Cardano transactions which are in the Cardano memory pool as ongoing
The problem that we didn't take into the considerations
transactions which have been created but haven't been submitted to the Cardano blockchain yet.
Because of that there existed gap between an object moved to AWAIT_CONFIRMATION state
and getWalletTransactions returning transaction for it.
This was considered weird so this gap is fixed by this commit.